Chore/#135: Helpers 수정
