+++ Harubee-iOS/Sources/Helpers/Manager/HapticManager.swift
