+++ Harubee-iOS/Sources/Presentation/Onboarding/Views/Onboarding3View.swift
@@ -7,37 +7,6 @@
 //
 
 import SwiftUI
-import UIKit
-
-extension UIApplication {
-  func endEditing() {
-    sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
-  }
-}
-
-@Observable
-final class KeyboardObserver {
-    var isKeyboardVisible: Bool = false
-
-    init() {
-        // 키보드가 나타나는 경우
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow), name: UIResponder.keyboardWillShowNotification, object: nil)
-        // 키보드가 사라지는 경우
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillHide), name: UIResponder.keyboardWillHideNotification, object: nil)
-    }
-
-    @objc private func keyboardWillShow() {
-        isKeyboardVisible = true
-    }
-
-    @objc private func keyboardWillHide() {
-        isKeyboardVisible = false
-    }
-
-    deinit {
-        NotificationCenter.default.removeObserver(self)
-    }
-}
 
 struct Onboarding3View: View {
   private var viewModel: OnboardingViewModel
@@ -46,8 +15,6 @@ struct Onboarding3View: View {
   @State private var isEnabled: Bool
   @State private var incomeDay: Int
   @State private var incomeAmount: String
-  
-  @State private var keyboardObserver = KeyboardObserver()
   
   init(viewModel: OnboardingViewModel) {
     self.viewModel = viewModel
@@ -63,8 +30,7 @@ struct Onboarding3View: View {
       
       OnboardingBodyView(
         incomeDay: $incomeDay,
-        incomeAmount: $incomeAmount,
-        isKeyboardVisible: $keyboardObserver.isKeyboardVisible
+        incomeAmount: $incomeAmount
       )
       
       Spacer()
@@ -122,16 +88,13 @@ private struct OnboardingBodyView: View {
   
   @Binding private var incomeDay: Int
   @Binding private var incomeAmount: String
-  @Binding private var isKeyboardVisible: Bool
   
   init(
     incomeDay: Binding<Int>,
-    incomeAmount: Binding<String>,
-    isKeyboardVisible: Binding<Bool>
+    incomeAmount: Binding<String>
   ) {
     self._incomeDay = incomeDay
     self._incomeAmount = incomeAmount
-    self._isKeyboardVisible = isKeyboardVisible
   }
   
   var body: some View {
@@ -139,8 +102,7 @@ private struct OnboardingBodyView: View {
       DayPickerView(
         title: "주요 수입일은 언제인가요?",
         titleFont: .onboarding,
-        selectedDay: $incomeDay,
-        isKeyboardVisible: $isKeyboardVisible
+        selectedDay: $incomeDay
       )
       .padding(.top, 34)
