+++ Harubee-iOS/Sources/Presentation/Setting/Views/FixedIncomeView.swift
@@ -17,8 +17,6 @@ struct FixedIncomeView: View {
   @State private var isInfoBubbleVisible: Bool = false
   @State private var isAlertPresented: Bool = false
   
-  @State private var keyboardObserver = KeyboardObserver()
-  
   private var settingViewModel: SettingViewModel
   
   init(
@@ -38,8 +36,7 @@ struct FixedIncomeView: View {
         BodyView(
           settingViewModel: settingViewModel,
           selectedDay: $selectedDay,
-          fixedIncomeAmount: $fixedIncomeAmount,
-          isKeyboardVisible: $keyboardObserver.isKeyboardVisible
+          fixedIncomeAmount: $fixedIncomeAmount
         )
         
         Spacer()
@@ -151,20 +148,17 @@ private struct BodyView: View {
   @State private var showingSheet: Bool = false
   @Binding private var selectedDay: Int
   @Binding private var fixedIncomeAmount: Int
-  @Binding private var isKeyboardVisible: Bool
   
   private var settingViewModel: SettingViewModel
   
   init(
     settingViewModel: SettingViewModel,
     selectedDay: Binding<Int>,
-    fixedIncomeAmount: Binding<Int>,
-    isKeyboardVisible: Binding<Bool>
+    fixedIncomeAmount: Binding<Int>
   ) {
     self.settingViewModel = settingViewModel
     self._selectedDay = selectedDay
     self._fixedIncomeAmount = fixedIncomeAmount
-    self._isKeyboardVisible = isKeyboardVisible
   }
   
   var body: some View {
@@ -172,8 +166,7 @@ private struct BodyView: View {
       DayPickerView(
         title: "주요 고정수입 날짜",
         titleFont: .view,
-        selectedDay: $selectedDay,
-        isKeyboardVisible: $isKeyboardVisible
+        selectedDay: $selectedDay
       )
       
       HStack(spacing: 0) {
